requirements.txt
streamlit
pandas
Pillow
duckduckgo_search
edge-tts
requests
streamlit-mic-recorder
openpyxl
streamlit-extras
python-docx
groq